PMM-7492 refactor form cleanup function
